refactor: 브릿지 코드 리팩토링
- pom.xml: 모드버스 의존성 추가 및 그룹 아이디 수정
- 프로젝트 패키지 이름 변경
- 프로퍼티 접두사 변경
- 데이터 DTO 접근제한자 변경
- 오탈자 수정
- 코드 스타일 정리
