+++ pom.xml
@@ -10,7 +10,7 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
-    <groupId>org.example</groupId>
+    <groupId>live.ioteatime</groupId>
     <artifactId>bridge-server</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>bridge-server</name>
@@ -48,6 +48,17 @@
             <groupId>com.fasterxml.jackson.datatype</groupId>
             <artifactId>jackson-datatype-jsr310</artifactId>
             <version>2.17.0</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.plc4x</groupId>
+            <artifactId>plc4j-api</artifactId>
+            <version>0.12.0</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.plc4x</groupId>
+            <artifactId>plc4j-driver-modbus</artifactId>
+            <version>0.12.0</version>
+            <scope>runtime</scope>
         </dependency>
     </dependencies>
