--- src/main/java/org/example/bridgeserver/domain/Data.java
@@ -1,18 +1,0 @@
-package org.example.bridgeserver.domain;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-@NoArgsConstructor
-@ToString
-@Getter
-public class Data {
-
-    @Setter
-    private String topic;
-    private Long time;
-    @Setter
-    private Float value;
-}
