+++ src/main/resources/application-dev.properties
@@ -7,10 +7,18 @@ spring.rabbitmq.password=guest
 spring.rabbitmq.template.exchange=amq.topic
 spring.rabbitmq.template.routing-key=data/#
 
+# mqtt, modbus
+bridge.server.protocol=mqtt
+
 mqtt.server.uri=tcp://133.186.153.19:1883
 mqtt.client.id=${random.uuid}
-mqtt.subscribe.topic=\
+mqtt.subscribe.topics=\
   data/s/nhnacademy/b/gyeongnam/p/class_a/d/gems-3500/e/electrical_energy/t/ac_indoor_unit/ph/1/de/power_factor,\
   data/s/nhnacademy/b/gyeongnam/p/class_a/d/gems-3500/e/electrical_energy/t/ac_indoor_unit/ph/1/de/v1
 
-bridge.mode=mqtt
+modbus.server.uri=tcp://192.168.70.220:502
+modbus.request.channels=\
+  4/8000/2,\
+  4/8002/2,\
+  4/8004/2,\
+  4/8006/2
